Add suse support for the snaps build
suse snaps rpm build and test happens inside a docker container
This patch requires the following patch to be merged:

https://git.opendaylight.org/gerrit/#/c/68114/

+++ jjb/packaging/stop-odl.sh
@@ -6,6 +6,21 @@
 #   -o pipefail: Fail on errors in scripts this calls, give stacktrace
 set -ex -o pipefail
 
+#-------------------------------------------------------------------------------
+# Exit if suse and in a VM
+#-------------------------------------------------------------------------------
+# Jenkins template and scripts are shared between suse and red hat to build and
+# test the rpms. However, all the suse processing is done in a container whereas
+# redhat processing is done in a VM. We should exit if we detect that this
+# script is going to test a suse rpm inside a VM. DISTRO variable only exists
+# when the script is executed in the VM.
+#-------------------------------------------------------------------------------
+if [ "$DISTRO" == "suse" ]; then
+  echo "We are in a VM, nothing to do for suse"
+  exit 0
+fi
+
+
 # Stop OpenDaylight
 sudo systemctl stop opendaylight
